+++ src/desktop2stereo/gui/gui.py
@@ -77,6 +77,7 @@ class Desktop2StereoGUI(
         self._calibration_run_requested = False
         self._calibration_poll_task = None
         self._calibration_dialog = None
+        self._display_refresh_warning_dialog = None
         self._calibration_active = False
         self._calibration_previous_target_value = None
         self.audio_devices = []